fix: ajustes na autenticacao
